♻️ 코드 수정: inbody의 height 타입 변경

+++ healthtart/src/main/java/com/dev5ops/healthtart/rival/domain/dto/RivalUserInbodyDTO.java
@@ -26,7 +26,7 @@ public class RivalUserInbodyDTO {
 
     // inbody 데이터
     private Integer inbodyScore;
-    private Integer height;
+    private Double height;
     private Double weight;
     private double muscleWeight;
     private double fatWeight;
